requirements.txt
urljoin
beautifulsoup4
patool
dbfread